improved backend makefile, added boolean support and unpack to Memory, added %lld support to BasicIO.printf

+++ test/struct.c
@@ -2,10 +2,10 @@
 
 typedef struct {
     unsigned char a;
-    int b;
+    long long b;
 } struct_t;
-struct_t s = {200,250};
+struct_t s = {200, 250};
 
 int main() {
-    printf("s = {%d, %d}\n", s.a, ++s.b);
+    printf("s = {%d, %lld}\n", s.a, ++s.b);
 }
